Refactored code (Removed unused awt.Shape from parser.Shape hierarchy, changed some constants from byte to int due to inconvenience in use and also did some other small changes)

+++ itext/src/main/java/com/itextpdf/text/pdf/parser/ExtRenderListener.java
@@ -31,5 +31,5 @@ public interface ExtRenderListener extends RenderListener {
      *
      * @param rule Either {@link PathPaintingRenderInfo#EVEN_ODD_RULE} or {@link PathPaintingRenderInfo#NONZERO_WINDING_RULE}
      */
-    void clipPath(byte rule);
+    void clipPath(int rule);
 }
